rework VerifySsoToken's form and resp
