+++ portal/web/api/v1/handlers/sso.go
@@ -3,16 +3,19 @@ package handlers
 import (
 	"cloudiac/portal/apps"
 	"cloudiac/portal/libs/ctx"
+	"cloudiac/portal/models/forms"
 )
 
 // TODO: swagger docs
+// GenerateSsoToken 生成 SSO token
 func GenerateSsoToken(c *ctx.GinRequest) {
 	c.JSONResult(apps.GenerateSsoToken(c.Service()))
 }
 
 // TODO: swagger docs
+// VerifySsoToken 验证 SSO token
 func VerifySsoToken(c *ctx.GinRequest) {
-	form := apps.VerifySsoTokenForm{}
+	form := forms.VerifySsoTokenForm{}
 	if err := c.Bind(&form); err != nil {
 		return
 	}
